requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Authentication and Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Background tasks
celery==5.3.4

# Validation and utilities
email-validator==2.1.0
python-dateutil==2.8.2

# Development dependencies (move to requirements-dev.txt in production)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0